Creación de componente desktopNavbar
