+++ src/components/navbar/desktop/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./desktopNavbar";
