+++ src/components/navbar/movile/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./desktopNavbar";
