+++ src/components/navbar/styles.ts
@@ -1,7 +1,6 @@
 export const navbarStyles = {
   background: "white",
   display: "flex",
-  justifyContent: "space-between",
   boxShadow: { xs: "none", sm: "0px 2px 2px 0px rgba(0, 0, 0, 0.15)" },
 };
 
@@ -19,4 +18,5 @@ export const titleStyles = {
 export const inlineStyles = {
   display: "flex",
   flexWrap: "wrap",
+  p: "15px",
 };
